+++ packages/orq-rc/RELEASES.md
@@ -7099,3 +7099,13 @@ Based on:
 - [typescript v4.4.0-rc.27] packages/orq-rc
 ### Releases
 - [NPM v4.4.0-rc.27] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.27 - packages/orq-rc
+
+## 2026-02-23 08:53:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.724.0 (2.836.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.4.0-rc.28] packages/orq-rc
+### Releases
+- [NPM v4.4.0-rc.28] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.28 - packages/orq-rc